requirements.txt
Django>=5,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5,<6
django-environ>=0.11,<1
celery>=5.4,<6
redis>=5,<6
pillow>=10,<11
pytest>=8,<9
pytest-django>=4,<5
drf-spectacular